stringManipulatins files
