+++ isaychev.vladislav/P3/addLat.hpp
@@ -1,13 +1,9 @@
 #ifndef ADDLAT_HPP
 #define ADDLAT_HPP
 
-#include <cstdlib>
-
 namespace isaychev
 {
   void addLat(const char * str1, const char * str2, char * str3);
-  void fillInUniqueChars(const char * str_base, char * str_final, size_t & counter);
-  void sortCharsAZ(char * str3);
 }
 
 #endif
