+++ isaychev.vladislav/P3/extractChars.hpp
@@ -1,12 +1,9 @@
 #ifndef EXTRACTCHARS_HPP
 #define EXTRACTCHARS_HPP
 
-#include <cstddef>
-
 namespace isaychev
 {
   void extractChars(const char * str_base, const char * str_sec, char * wk_str);
-  size_t checkForSameElem(const char * str2, char charForCheck);
 }
 
 #endif
